runtime: Fix string/array type definition
Currently compiler uses int32 for data size; using size_t can lead to reading
uninitialized data.

+++ runtime/types.hpp
@@ -5,13 +5,13 @@ union TypeInfo;
 struct AikeString
 {
 	const char* data;
-	size_t size;
+	int size;
 };
 
 template <typename T> struct AikeArray
 {
 	T* data;
-	size_t size;
+	int size;
 };
 
 struct AikeAny
